+++ src/assets.ts
@@ -67,6 +67,9 @@ export function assetToName(asset: Asset): string | null {
   if (asset == Asset.RNDR) return "RNDR";
   if (asset == Asset.TNSR) return "TNSR";
   if (asset == Asset.POPCAT) return "POPCAT";
+  if (asset == Asset.EIGEN) return "EIGEN";
+  if (asset == Asset.DBR) return "DBR";
+  if (asset == Asset.GOAT) return "GOAT";
   if (asset == null) return null; // Some things, like clock callbacks, are for all assets and return asset=null
   return "UNDEFINED";
 }
@@ -90,6 +93,9 @@ export function nameToAsset(name: string): Asset {
   if (name == "RNDR") return Asset.RNDR;
   if (name == "TNSR") return Asset.TNSR;
   if (name == "POPCAT") return Asset.POPCAT;
+  if (name == "EIGEN") return Asset.EIGEN;
+  if (name == "DBR") return Asset.DBR;
+  if (name == "GOAT") return Asset.GOAT;
   return Asset.UNDEFINED;
 }
 
@@ -116,6 +122,9 @@ export function toProgramAsset(asset: Asset): any {
   if (asset == Asset.RNDR) return { rndr: {} };
   if (asset == Asset.TNSR) return { tnsr: {} };
   if (asset == Asset.POPCAT) return { popcat: {} };
+  if (asset == Asset.EIGEN) return { eigen: {} };
+  if (asset == Asset.DBR) return { dbr: {} };
+  if (asset == Asset.GOAT) return { goat: {} };
   return { undefined: {} };
 }
 
@@ -173,6 +182,15 @@ export function fromProgramAsset(asset: any): Asset {
   }
   if (objectEquals(asset, { popcat: {} })) {
     return Asset.POPCAT;
+  }
+  if (objectEquals(asset, { eigen: {} })) {
+    return Asset.EIGEN;
+  }
+  if (objectEquals(asset, { dbr: {} })) {
+    return Asset.DBR;
+  }
+  if (objectEquals(asset, { goat: {} })) {
+    return Asset.GOAT;
   }
   return Asset.UNDEFINED;
 }
@@ -233,6 +251,15 @@ export function assetToIndex(asset: Asset): number {
     case Asset.POPCAT: {
       return 17;
     }
+    case Asset.EIGEN: {
+      return 18;
+    }
+    case Asset.DBR: {
+      return 19;
+    }
+    case Asset.GOAT: {
+      return 20;
+    }
   }
   return 255; // Undefined is 255 onchain
 }
@@ -292,6 +319,15 @@ export function indexToAsset(index: number): Asset {
     }
     case 17: {
       return Asset.POPCAT;
+    }
+    case 18: {
+      return Asset.EIGEN;
+    }
+    case 19: {
+      return Asset.DBR;
+    }
+    case 20: {
+      return Asset.GOAT;
     }
   }
   return Asset.UNDEFINED;